Rename "stock" widget to "reporting rates"
Issue: #2574
