+++ static/css/analytics-widgets.less
@@ -148,7 +148,7 @@
   color: #1D86F0;
 }
 
-.stock-monitoring {
+.reporting-rates {
   padding: 0.5em;
 
   .row {
