+++ static/js/modules/reporting-rates-utils.js
