+++ translations/messages-en.properties
@@ -212,10 +212,10 @@ analytics.anc.visits-completed = Visits Completed During Pregnancy
 analytics.anc.visits-completed.description = Completed pregnancies that have had...
 analytics.anc.visits-during = Visits Completed So Far
 analytics.anc.visits-during.description = Active pregnancies that have had...
-analytics.stock.health_centers_size = SMO
-analytics.stock.clinics_size = Reporting Unit
-analytics.stock.not_submitted = Reports Missing
-analytics.stock.incomplete = Invalid Reports
+analytics.reporting.health_centers_size = SMO
+analytics.reporting.clinics_size = Reporting Unit
+analytics.reporting.not_submitted = Reports Missing
+analytics.reporting.incomplete = Invalid Reports
 Previous = Prev
 Next = Next
 End\ tour = End tour
@@ -274,7 +274,7 @@ Error\ saving\ feedback = Error saving feedback
 feedback.submitted = Feedback submitted
 Error\ updating\ group = Error updating group
 Antenatal\ Care = Antenatal Care
-Stock\ Monitoring = Stock Monitoring
+Reporting\ Rates = Reporting Rates
 Number\ in\ month = {{count}} in {{month}}
 Schedules = Schedules
 Users = Users
