+++ translations/messages-es.properties
@@ -308,7 +308,6 @@ Please\ select\ a\ facility = Por favor seleccionar un contacto
 Error\ saving\ feedback = Error guardando comentarios
 Error\ updating\ group = Error actualizando grupo
 Antenatal\ Care = Atención prenatal
-Stock\ Monitoring = Monitoreo de reservas
 Number\ in\ month = {{count}} en {{month}}
 # Depending on the type of schedule this refers to, "programas" might be a better alternative to "horario".
 # Fuzzy
