+++ translations/messages-fr.properties
@@ -266,7 +266,6 @@ Please\ select\ a\ facility = S'il vous plaît sélectionnez un contact
 Error\ saving\ feedback = Erreur lors de l'enregistrement du commentaire
 Error\ updating\ group = Erreur lors de la mise à jour du groupe
 Antenatal\ Care = Soins prénatals
-Stock\ Monitoring = Suivi des stocks
 Number\ in\ month = {{count}} en {{month}}
 Schedules = Horaires
 Users = Utilisateurs
