+++ translations/messages-hi.properties
@@ -302,7 +302,6 @@ Error\ saving\ feedback = सुझावों को बचाने में
 Error\ updating\ group = समूह के नवीकरण में त्रुटि
 # Translated as "pregnancy care"
 Antenatal\ Care = गर्भावस्था की देखभाल
-Stock\ Monitoring = स्टॉक की निगरानी
 Number\ in\ month = {{count}} में {{month}}
 Schedules = निर्धारित
 Users = उपयोगकर्ताओं
