+++ translations/messages-ne.properties
@@ -267,7 +267,6 @@ Please\ select\ a\ facility = कृपया संस्था रोज्न
 Error\ saving\ feedback = प्रतिक्रिया राख्नमा त्रुटी   
 Error\ updating\ group = समूह परिमार्जन गर्नेमा त्रुटी 
 Antenatal\ Care = पुर्व प्रसूति सेवा
-Stock\ Monitoring = आपूर्ति अनुगमन
 Number\ in\ month = महिना मा गणना
 Schedules = तालिकाहरु
 Users = प्रयोगकर्ता
