+++ translations/messages-sw.properties
@@ -268,7 +268,6 @@ Please\ select\ a\ facility = tafadhali chagua mwasiliani
 Error\ saving\ feedback = kuna  makosa  kusajili  maoni
 Error\ updating\ group = kuna makosa kubadilisha  kundi
 Antenatal\ Care = huduma wakati mama ana ujauzito
-Stock\ Monitoring = ufuatiliaji hisa
 Number\ in\ month = idadi iliyoko  mwezini
 Schedules = ratiba
 Users = watumiaji
